# sutra/contracts/sutra.openapi.yaml
openapi: 3.0.3
info:
  title: sutra
  description: The single API surface — CLI, web UI, and external agents all speak this contract.
  version: 0.2.0
paths:
...
  /projects:
    post:
      operationId: createProject
      summary: Create a project
      description: >-
        Creates a project. Used directly by `sutra init` (AC-project-init), which then
        records the repo path here and writes a local marker linking the directory to
        the returned project id.
      tags: [projects]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewProject'
      responses:
        '201':
          description: Project created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listProjects
      summary: List projects
      description: Archived projects are excluded unless includeArchived=true (AC-project-archive).
      tags: [projects]
      parameters:
        - name: includeArchived
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Projects.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Project' }
  /projects/{projectId}:
    get:
      operationId: getProject
      summary: Get a project
      tags: [projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
      responses:
        '200':
          description: The project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/archive:
    post:
      operationId: archiveProject
      summary: Archive a project
      description: >-
        Hides the project from default listings and makes it read-only; nothing is
        deleted (AC-project-archive).
      tags: [projects]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actor]
              properties:
                actor: { type: string, format: uuid }
      responses:
        '200':
          description: Project archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The project is already archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /projects/{projectId}/issues:
    post:
      operationId: createIssue
      summary: Create an issue
      description: >-
        Mints a UUIDv7 id plus a per-project sequential display number, sets status
        open, and records a creation event (AC-issue-create).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIssue'
      responses:
        '201':
          description: Issue created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listIssues
      summary: List and search issues in a project
      description: >-
        Filters by status, assignee, and label, combinable with each other and with
        text search over title/body/comments (AC-search-text, AC-search-filters).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/ProjectIdPath'
        - name: number
          in: query
          description: >-
            Filter to the issue with this project-scoped display number
            (e.g. 42 for SUT-42), so /p/:key/i/:num and `sutra issue show
            SUT-42` can resolve directly instead of listing and scanning.
          schema: { type: integer }
        - name: status
          in: query
          style: form
          explode: true
          schema:
            type: array
            items: { $ref: '#/components/schemas/IssueStatus' }
        - name: assignee
          in: query
          schema: { type: string, format: uuid }
        - name: label
          in: query
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
        - name: q
          in: query
          description: Free-text search over issue titles, bodies, and comments.
          schema: { type: string }
      responses:
        '200':
          description: Matching issues, ranked by relevance when q is set.
          content:
            application/json:
              schema:
                type: object
                required: [issues, feed_watermark]
                properties:
                  issues:
                    type: array
                    items: { $ref: '#/components/schemas/Issue' }
                  feed_watermark:
                    type: string
                    description: Server-issued feed position captured atomically with this query — present even when issues is empty, so a consumer can always reconcile the feed against exactly what this read reflected.
  /issues/{issueId}:
    get:
      operationId: getIssue
      summary: Get an issue
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
      responses:
        '200':
          description: The issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueRead'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateIssue
      summary: Update an issue's title, body, or fields
      description: Changes persist and `updated` reflects the mutation (AC-issue-update).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssue'
      responses:
        '200':
          description: Updated issue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueRead'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /issues/{issueId}/status:
    post:
      operationId: updateIssueStatus
      summary: Transition an issue's status
      description: >-
        Transitions among open, in-progress, blocked, and deferred are free and
        recorded as events; complete is gated on an approved Review NAMED in
        the request (`review`, required exactly when status is complete): it
        must belong to the issue, be approved at its current revision, and
        not be close-used; the successful transition atomically stamps it
        close-used and freezes its verdict (AC-close-approved) — a
        merge-time consumption never blocks the close, a spent review never
        authorizes another one, and a reopened issue closes again only under
        a fresh approval — plus a DESCENDANT-AWARE check — rejected while any descendant at any depth,
        including beneath deferred children, is open, in-progress, or blocked
        (AC-status-set, AC-status-reopen, AC-parent-close-gate,
        AC-close-approved, AC-close-blocked-without). An issue entering open,
        in-progress, or blocked atomically reopens every complete ancestor in
        the same transaction, with a status event recorded on each
        (AC-parent-reopen-cascade); every transition transaction also increments
        subtree_revision by exactly one on each affected issue and its
        ancestors — once per transaction, however many cascaded transitions
        it produces (AC-subtree-revision).
        Invoked by the kanban board's
        drag-to-move action (ACT-move-card). The request may carry
        expected_status: when present and the issue's current status
        differs, the transition is rejected with a conflict — optimistic
        concurrency for racing writers, e.g. a deferral racing a
        work-stack pop (AC-status-conditional). A complete transition may
        additionally carry expected_subtree_revision: mismatch rejects with a
        conflict even when current state matches, fencing histories where a
        descendant reopened and recompleted since the caller's observation.
        Replay semantics matter especially here (CON-idempotent-mutations):
        a replayed successful close returns its original success — never a
        close_used conflict, because close_used was stamped by THIS key's
        original execution — and a replayed reopen returns the original
        response with no second status event, cascade, or subtree_revision
        increment.
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - title: CompleteTransition
                  type: object
                  required: [status, review, review_revision, review_verdict_event, actor]
                  properties:
                    status: { type: string, enum: [complete] }
                    expected_status: { $ref: '#/components/schemas/IssueStatus' }
                    expected_subtree_revision:
                      type: integer
                      description: Rejected with a conflict if the issue's subtree_revision differs — a history-aware fence.
                    review:
                      type: string
                      format: uuid
                      description: The authorizing review — must belong to the issue, be approved at its current revision, and not be close-used.
                    review_revision:
                      type: integer
                      description: The approved revision the caller expects to spend; validated atomically with ownership, approval state, and close_used — a delayed close intended for revision N can never spend a later revision's approval.
                    review_verdict_event:
                      type: string
                      format: uuid
                      description: The approval's verdict event; must still be the review's latest — a verdict ABA rejects the delayed close atomically.
                    actor: { type: string, format: uuid }
                  additionalProperties: false
                - title: NonCompleteTransition
                  type: object
                  required: [status, actor]
                  properties:
                    status: { type: string, enum: [open, in-progress, blocked, deferred] }
                    expected_status: { $ref: '#/components/schemas/IssueStatus' }
                    actor: { type: string, format: uuid }
                  additionalProperties: false
      responses:
        '200':
          description: Issue transitioned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            The transition is rejected: closing with a NAMED review that does
            not belong to the issue, is not approved at its current revision,
            or is already close-used (omitting review entirely on a complete
            transition fails schema validation as 400, never 409); an active
            descendant at any depth — including beneath deferred children —
            holding the issue open; or the issue's project being archived.
            An expected_status precondition that does not match the issue's
            current status also conflicts (AC-status-conditional), as does an
            expected_subtree_revision that does not match the issue's current
            subtree_revision (AC-subtree-revision) — history-aware, so a
            descendant that reopened and recompleted still conflicts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /issues/{issueId}/assign:
    post:
      operationId: assignIssue
      summary: Assign or unassign an issue
      description: >-
        Assigns to any identity, human or agent, or clears assignment with an
        explicit null assignee; assignment feeds the agent work stack
        (AC-issue-assign). assignee is required in the request — an absent
        assignee is invalid; to clear assignment, send assignee: null.
        Clearing the assignee emits an `issue.unassigned` event.
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignee, actor]
              properties:
                assignee:
                  type: string
                  format: uuid
                  nullable: true
                  description: >-
                    Required. Set to an identity id to assign, or explicit
                    null to clear; absent is invalid.
                actor: { type: string, format: uuid }
      responses:
        '200':
          description: Issue reassigned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueRead'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /issues/{issueId}/relations:
    post:
      operationId: addIssueRelation
      summary: Add a parent/child or blocking relation
      description: >-
        Creates a parent_of or blocks relation from this issue. Ancestry cycles
        (AC-parent-create) and blocking cycles, direct or transitive
        (AC-block-no-cycles), are rejected. Conflict codes are NORMATIVE for
        this operation: an already-present relation returns relation-exists,
        an ancestry cycle ancestry-cycle, a blocking cycle blocking-cycle,
        and an archived project project-archived — no other code is
        contract-valid for those conditions (AC-relation-conflict-codes). Every parent_of attachment increments subtree_revision on
        the new parent and each of its ancestors, whether or not any
        reopening occurs (AC-subtree-revision). An attachment that places
        active work beneath a complete parent — the attached child itself, or
        any descendant of its subtree, being open, in-progress, or blocked —
        additionally reopens every complete ancestor atomically in the same
        transaction, recording a status event on each
        (AC-parent-reopen-cascade). Records an
        `issue.relation-added` event (subject: the relation's `from` issue).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewIssueRelation'
      responses:
        '201':
          description: Relation created.
          content:
            application/json:
              schema:
                type: object
                required: [relation, operation]
                properties:
                  relation: { $ref: '#/components/schemas/IssueRelation' }
                  operation:
                    type: string
                    format: uuid
                    description: The server transaction id this mutation ran under — the same id every event it emitted carries, cascades included, so the caller (and its idempotent replay, which returns this original response) can correlate the attachment's cascade directly.
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: >-
            The relation would create a cycle, this issue already has a parent,
            or the issue's project is archived.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelationConflictError'
    get:
      operationId: listIssueRelations
      summary: List an issue's relations
      description: >-
        Backs parent/child views (parent lists children, child shows parent,
        AC-parent-create) and blocking views (AC-block-create) from both sides.
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
      responses:
        '200':
          description: Relations where this issue is either side.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IssueRelation' }
  /issues/{issueId}/relations/{relationId}:
    delete:
      operationId: removeIssueRelation
      summary: Remove a relation
      description: >-
        Removal clears both sides (AC-block-create) and records an
        `issue.relation-removed` event whose payload snapshots the removed
        relation — its id, kind, former `from`, and former `to` — so a
        consumer can distinguish a parent_of detachment from a removed
        blocks relation and identify the detached child. Removing a parent_of relation
        atomically increments subtree_revision on the former parent and each
        of its ancestors (AC-subtree-revision) — a detached-then-reopened
        child cannot leave a stale revision satisfying a close fence.
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
        - $ref: '#/components/parameters/RelationIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: actor
          in: query
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: Relation removed.
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /issues/{issueId}/events:
    get:
      operationId: listIssueEvents
      summary: Get an issue's audit history
      description: >-
        All events for the issue, in chronological order, with actor, change kind,
        and time (AC-audit-query).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
      responses:
        '200':
          description: Audit history.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Event' }
  /issues/{issueId}/documents:
    get:
      operationId: listIssueDocuments
      summary: List documents currently tied to an issue
      description: An issue lists its currently tied documents (AC-doc-list-by-issue).
      tags: [issues]
      parameters:
        - $ref: '#/components/parameters/IssueIdPath'
      responses:
        '200':
          description: Tied documents.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }
...
  /identities/{identityId}/work-stack/pop:
    post:
      operationId: popWorkStack
      summary: Atomically claim the next workable issue assigned to an identity
      description: >-
        Returns the oldest workable assigned issue and sets its status to
        in-progress in the same operation (AC-pop-claims, AC-pop-fifo). Concurrent
        pops by the same identity never return the same issue twice
        (AC-pop-atomic). A blocked issue whose blocker is also assigned to this
        identity is walked to the deepest open blocker (AC-pop-blocker-first); one
        blocked by an issue assigned elsewhere is skipped (AC-pop-skips-blocked).
        With no workable issues, returns an explicit empty result, not an error
        (AC-pop-empty). Idempotent (CON-idempotent-mutations): replaying a pop
        with the same Idempotency-Key returns the same claimed issue again —
        it never claims a second issue.
        Issues belonging to archived projects are skipped like externally
        blocked issues — archived projects are read-only, so their work is
        never handed out. Eligibility means status open: the judgment
        statuses blocked and deferred are never handed out, nor are
        in-progress, complete, or relation-blocked issues. The pop-induced `issue.status-changed` event's
        actor is the popping identity.
      tags: [work-stack]
      parameters:
        - $ref: '#/components/parameters/IdentityIdPath'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      responses:
        '200':
          description: The claimed issue, or an empty result if none was workable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkStackPopResult'
        '404':
          $ref: '#/components/responses/NotFound'

  # ------------------------------------------------------------------ labels
...
  /reviews/{reviewId}:
    get:
      operationId: getReview
      summary: Get a review
      description: Shows the deliverable's content for reading (AC-review-web).
      tags: [reviews]
      parameters:
        - $ref: '#/components/parameters/ReviewIdPath'
      responses:
        '200':
          description: The review.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Client-supplied key that makes this mutating operation safe to retry
        (CON-idempotent-mutations): replaying the same key returns the
        original response verbatim and has no second effect — no duplicate
        resource, no second state transition, no second side effect.
      schema: { type: string }
    ProjectIdPath:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: >-
        The request is invalid — e.g. an actor or assignee identity id that
        names no known identity (AC-identity-referenced), or a status outside
        the five defined values (AC-status-invalid). The API deals only in
        identity ids; resolving a human-typed handle to an id is the CLI's job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    NewProject:
      type: object
      required: [key, name, actor]
      properties:
        key: { type: string }
        name: { type: string }
        description: { type: string }
        repo_path: { type: string }
        default_branch:
          type: string
          description: >-
            Diff base for code deliverables; defaults to "main" when omitted.
        actor:
          type: string
          format: uuid
          description: Identity recorded as the actor on the creation event.
...